'accomplished'

+++ src/java/servlet/DeleteNotes.java
@@ -15,29 +15,35 @@ import java.sql.*;
  */
 public class DeleteNotes {
 
-    public String dn(Connection con, String Note) {
+    public String dn(Connection con, String Note,String UserId) {
         PreparedStatement ps;
+        char perm='\0';
         String msg = "";
         ResultSet rs = null;
         try {
-            ps = con.prepareStatement("select * from notes where id=" + Note); //+ " and name='" + Note + "'");
+            ps = con.prepareStatement("select permission from registry where id=" + Note +" and userid="+UserId); //+ " and name='" + Note + "'");
             rs = ps.executeQuery();
             String check = "";
             try {
-                rs.next();
-                check = rs.getString(1);
-                
+               if(rs.next())
+               {
+                   perm=rs.getString(1).toCharArray()[0];
+               }
             } catch (Exception e) {
                 msg = "no note with id = " + Note + " is available";
                 return msg;
             }
-             ps = con.prepareStatement("delete from registry where id=" + Note );//+ " and name='" + Note + "'");
+            if(perm=='r')
+            {
+                return "no write permissions granted";
+            }
+            ps = con.prepareStatement("delete from registry where id=" + Note);//+ " and name='" + Note + "'");
             ps.executeUpdate();
-            ps = con.prepareStatement("delete from notes where id=" + Note );//+ " and name='" + Note + "'");
+            ps = con.prepareStatement("delete from notes where id=" + Note);//+ " and name='" + Note + "'");
             ps.executeUpdate();
             msg = "note with id= '" + Note + "' is deleted";
         } catch (Exception e) {
-            msg=e.toString();
+            msg = e.toString();
         }
         return msg;
     }
